feat(soup): share email threads from the context menu
Email threads were the only shareable entity missing a Share item in the
soup context menu (and its hotkey): `isShareableEntityType` covered only
document, chat and project, so `makeShareAction.canExecute` returned false
for email rows and the item was never pushed.

Add `email` to the shareable types, gated on ENABLE_EMAIL_SHARING — the
same flag that gates the email block's Share tool. The share modal already
handles `email`: it forwards the thread to a channel via `editThread` and
hides link sharing/access-level selection for that item type.

Move the type and its guard into `shareable-entity.ts` so callers that only
need the predicate don't pull in the modal (and its websocket clients),
which also makes it unit-testable.
