+++ apps/web/src/features/sharing/global-share-modal/GlobalShareModal.tsx
@@ -3,16 +3,12 @@ import { Permissions } from '@core/component/SharePermissions';
 import { ShareModal } from '@core/component/TopBar/ShareButton';
 import { itemToBlockName } from '@core/constant/allBlocks';
 import { createControlledOpenSignal } from '@core/util/createControlledOpenSignal';
-import type { EntityData } from '@entity';
 import type { ItemType } from '@service-storage/client';
 import { createSignal, Show } from 'solid-js';
-
-type ShareableEntityType = 'document' | 'chat' | 'project';
-
-export type ShareableEntityData = Extract<
-  EntityData,
-  { type: ShareableEntityType }
->;
+import {
+  isShareableEntityType,
+  type ShareableEntityData,
+} from './shareable-entity';
 
 type GlobalShareModalProps = {
   entity: ShareableEntityData;
@@ -24,12 +20,6 @@ const [globalModalProps, setGlobalModalProps] =
 const [modalOpen, setModalOpen] = createControlledOpenSignal(false, {
   id: 'global-share',
 });
-
-export const isShareableEntityType = (
-  type: EntityData['type']
-): type is ShareableEntityType => {
-  return type === 'document' || type === 'chat' || type === 'project';
-};
 
 const getEntityBlockAlias = (
   entity: ShareableEntityData
